# Conflicts:
#	android_app/app/build.gradle
#	android_app/app/src/main/java/com/health/openscale/core/bluetooth/BluetoothQNScale.java
#	android_app/app/src/main/java/com/health/openscale/gui/preferences/BluetoothPreferences.java
#	android_app/app/src/main/res/values-de/strings.xml

+++ android_app/app/src/main/java/com/health/openscale/core/bluetooth/BluetoothOneByone.java
@@ -143,7 +143,7 @@ public class BluetoothOneByone extends BluetoothCommunication {
                 break;
         }
 
-        OneByoneLib oneByoneLib = new OneByoneLib(sex, scaleUser.getAge(), (int)scaleUser.getBodyHeight(), peopleType);
+        OneByoneLib oneByoneLib = new OneByoneLib(sex, scaleUser.getAge(), scaleUser.getBodyHeight(), peopleType);
 
         ScaleMeasurement scaleBtData = new ScaleMeasurement();
         scaleBtData.setWeight(weight);
